requirements.txt
docker
gunicorn
Flask
Flask-Testing
pytest
pytest-flask
pre-commit
pyyaml
